backend/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
jinja2==3.*
pydantic==2.*
google-cloud-storage==2.*
python-dotenv==1.*
